api/requirements.txt
django
django-ckeditor
djangorestframework
Pillow
django-cors-headers